requirements.txt
flask
requests
python-dotenv
dnspython
python-whois
ping3
speedtest-cli